fix(generate-registry): fix duplicated testdata
